+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 
 - support for grouping operationId by "group"
 - support for using Plug only, with no Phoenix dependency
+- support for aliasing operationId
 
 ## Future planned support